rename extensions to flask_extensions
